// src/frontend/src/backend.d.ts
export type Time = bigint;
export interface OrderRecord {
    id: string;
    status: OrderStatus;
    deliveryAddress: DeliveryAddress;
    paymentMethod: PaymentMethod;
    user: Principal;
    totalAmount: bigint;
    timestamp: Time;
    items: Array<OrderItem>;
}
export interface OrderItem {
    productId: string;
    quantity: bigint;
    priceAtOrder: bigint;
}
// ...
export interface DeliveryAddress {
    city: string;
    fullName: string;
    state: string;
    addressLine1: string;
    addressLine2: string;
    phone: string;
    pincode: string;
}
export interface Product {
    id: string;
    stockQuantity: bigint;
    isOnSale: boolean;
    name: string;
    description: string;
    imageUrl: string;
    category: string;
    salePrice?: bigint;
    price: bigint;
}
export interface UserProfile {
    name: string;
}
export enum IdempotenceResult {
    success = "success",
    alreadyInitialized = "alreadyInitialized"
}
export enum OrderStatus {
    shipped = "shipped",
    pending = "pending",
    paid = "paid",
    delivered = "delivered"
}
export enum PaymentMethod {
    cod = "cod",
    upi = "upi"
}
export enum UserRole {
    admin = "admin",
    user = "user",
    guest = "guest"
}